Moved app infrastructure to a allosphere module
