--- allosphere/src/al_AlloApp.cpp
@@ -1,2 +1,0 @@
-#include "allosphere/al_AlloApp.hpp"
-
