+++ cmake_modules/CMakeRunTargets.cmake
@@ -228,8 +228,8 @@ if(TARGET allosphere${DEBUG_SUFFIX})
 	get_target_property(ALLOSPHERE_DEP_INCLUDE_DIR allosphere${DEBUG_SUFFIX} ALLOSPHERE_DEP_INCLUDE_DIR)
 	get_target_property(ALLOSPHERE_INCLUDE_DIR allosphere${DEBUG_SUFFIX} ALLOSPHERE_INCLUDE_DIR)
 	get_target_property(ALLOSPHERE_LINK_LIBRARIES "allosphere${DEBUG_SUFFIX}" ALLOSPHERE_LINK_LIBRARIES)
-	add_dependencies("${APP_NAME}" allosphere${DEBUG_SUFFIX})
-	target_link_libraries("${APP_NAME}" ${ALLOSPHERE_LIBRARY} ${ALLOSPHERE_LINK_LIBRARIES})
+	add_dependencies("${ALLO_APP_NAME_SUFFIX}" allosphere${DEBUG_SUFFIX})
+	target_link_libraries("${ALLO_APP_NAME_SUFFIX}" ${ALLOSPHERE_LIBRARY} ${ALLOSPHERE_LINK_LIBRARIES})
 	include_directories(${ALLOSPHERE_INCLUDE_DIR} ${ALLOSPHERE_DEP_INCLUDE_DIR})
 else()
   if(NOT ALLOSPHERE_FOUND)
